# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: ProspectSA - Saudi B2B Intelligence Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /leads:
    get:
      operationId: listLeads
      tags: [leads]
      summary: List all saved leads
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: companyId
          in: query
          schema:
            type: ["integer", "null"]
        - name: search
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Lead list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeadListResponse"
    post:
      operationId: createLead
      tags: [leads]
      summary: Create a lead manually
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLeadBody"
      responses:
        "201":
          description: Created lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
  /leads/{id}:
    patch:
      operationId: updateLead
      tags: [leads]
      summary: Update a lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLeadBody"
      responses:
        "200":
          description: Updated lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
    delete:
      operationId: deleteLead
      tags: [leads]
      summary: Delete a lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ── MeshBase Executives ───────────────────────────────────────────────────────
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
...
    Company:
      type: object
      properties:
        id:
          type: integer
        nameAr:
          type: ["string", "null"]
        nameEn:
          type: ["string", "null"]
        industry:
          type: ["string", "null"]
        industryAr:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        region:
          type: ["string", "null"]
        country:
          type: string
          default: "Saudi Arabia"
        website:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        descriptionAr:
          type: ["string", "null"]
        employeeCount:
          type: ["integer", "null"]
        revenue:
          type: ["string", "null"]
        foundingYear:
          type: ["integer", "null"]
        crNumber:
          type: ["string", "null"]
        capitalAmount:
          type: ["string", "null"]
        entityType:
          type: ["string", "null"]
        companyType:
          type: ["string", "null"]
        ownerName:
          type: ["string", "null"]
        ownerNameAr:
          type: ["string", "null"]
        ownerTitle:
          type: ["string", "null"]
        ownerPhone:
          type: ["string", "null"]
        ownerEmail:
          type: ["string", "null"]
        ownerLinkedin:
          type: ["string", "null"]
        estimatedWealth:
          type: ["string", "null"]
        shareholders:
          type: ["string", "null"]
          description: JSON array of shareholders with ownership percentage
        keyExecutives:
          type: ["string", "null"]
          description: JSON array of key executives
        marketPositioning:
          type: ["string", "null"]
        recentNews:
          type: ["string", "null"]
          description: JSON array of recent news items
        linkedinUrl:
          type: ["string", "null"]
        twitterUrl:
          type: ["string", "null"]
        enrichmentScore:
          type: ["integer", "null"]
          description: 0-100 enrichment completeness score
        enrichmentStatus:
          type: ["string", "null"]
          enum: ["enriched", "partial", "pending", null]
        dataSource:
          type: ["string", "null"]
        tags:
          type: ["string", "null"]
          description: Comma-separated tags
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - country
        - createdAt
        - updatedAt
...
    Lead:
      type: object
      properties:
        id:
          type: integer
        companyId:
          type: ["integer", "null"]
        firstName:
          type: ["string", "null"]
        lastName:
          type: ["string", "null"]
        firstNameAr:
          type: ["string", "null"]
        lastNameAr:
          type: ["string", "null"]
        title:
          type: ["string", "null"]
        titleAr:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        linkedinUrl:
          type: ["string", "null"]
        twitterUrl:
          type: ["string", "null"]
        department:
          type: ["string", "null"]
        seniority:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        status:
          type: string
          enum: [new, contacted, qualified, converted, disqualified]
          default: new
        company:
          $ref: "#/components/schemas/Company"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - status
        - createdAt
        - updatedAt
    LeadListResponse:
      type: object
      properties:
        leads:
          type: array
          items:
            $ref: "#/components/schemas/Lead"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required:
        - leads
        - total
        - page
        - limit
        - totalPages
    CreateLeadBody:
      type: object
      properties:
        companyId:
          type: ["integer", "null"]
        firstName:
          type: ["string", "null"]
        lastName:
          type: ["string", "null"]
        title:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        linkedinUrl:
          type: ["string", "null"]
        department:
          type: ["string", "null"]
        seniority:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
    UpdateLeadBody:
      type: object
      properties:
        firstName:
          type: ["string", "null"]
        lastName:
          type: ["string", "null"]
        title:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        status:
          type: ["string", "null"]
          enum: [new, contacted, qualified, converted, disqualified, null]

    # ── Dashboard ─────────────────────────────────────────────────────────────
...